+++ changelog.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.0] - 2025-06-17
+### Added
+- `Lemma.prototype.getDeclension`
+- `Lemma.prototype.getSchoolDeclension`
+
+### Deprecated
+- `RussianNouns.getDeclension`
+- `RussianNouns.getSchoolDeclension`
+
+### Changed
+- Slight improvement in the declension of masculine possessive adjectives (not tested properly).
+
 ## [2.2.0] - 2025-04-16
 ### Changed
 - Improved declension of adjectives.